+++ lib/kitchen/directions/bake_further_research.rb
@@ -15,7 +15,7 @@ module Kitchen
       class V1
         renderable
         def bake(chapter:, metadata_source:, uuid_prefix: '.')
-          BakeGenericEocSection.v1(
+          MoveCustomSectionToEocContainer.v1(
             chapter: chapter,
             metadata_source: metadata_source,
             klass: 'further-research',