+++ spec/directions/move_custom_section_to_eoc_container/main_spec.rb
@@ -2,9 +2,9 @@
 
 require 'spec_helper'
 
-RSpec.describe Kitchen::Directions::BakeGenericEocSection do
+RSpec.describe Kitchen::Directions::MoveCustomSectionToEocContainer do
   it 'calls v1' do
-    expect_any_instance_of(Kitchen::Directions::BakeGenericEocSection::V1).to receive(:bake)
+    expect_any_instance_of(Kitchen::Directions::MoveCustomSectionToEocContainer::V1).to receive(:bake)
       .with(chapter: 'chapter1', metadata_source: 'metadata', append_to: 'element', klass: 'class', uuid_prefix: 'uuid', include_intro_page: true)
     described_class.v1(chapter: 'chapter1', metadata_source: 'metadata', append_to: 'element', klass: 'class', uuid_prefix: 'uuid')
   end